refactor: modularize diagnostics
