+++ main.py
@@ -46,13 +46,14 @@ from core.functions import logit, relu, cross_entropy, activation_dict, cost_dic
 from ui.dashboard_panel.cost.cost_view import CostWidget
 from ui.dashboard_panel.accuracy.accuracy_view import AccuracyWidget
 from ui.abstract.base_plot import PLOT_COST_HEIGHT, PLOT_COST_WIDTH
+from ui.abstract.button import Button
+from ui.diagnostics_panel.predict_training_set.predict_training_set_view import ImageTrainWidget
+from ui.diagnostics_panel.predict_validation_set.predict_validation_set_view import ImageValidationWidget
+from ui.diagnostics_panel.predict_test_set.predict_test_set_view import ImageTestWidget
+from ui.diagnostics_panel.nn_internal_state.matrix_view import FirstLayerWeightsWidget, FirstLayerGradientWidget
 
-PLOT_TEST_WIDTH = 650
-PLOT_TEST_HEIGHT = 650
 CONTROL_PANEL_WIDTH = 500
 CONTROL_PANEL_HEIGHT = 550
-WEIGHT_FILTER_WIDTH = 150
-WEIGHT_FILTER_HEIGHT = 150
 
 class IntegerChoiceWidget(QWidget):
 
@@ -228,28 +229,6 @@ class RunIdWidget(QWidget):
         layout.addWidget(label)
         layout.addWidget(self.input_text)
         self.setLayout(layout)
-
-class Button(QPushButton):
-    def __init__(self, label, run, parent=None, enabled=True):
-        super().__init__(parent)
-        self.setText(label)
-        self.setFixedHeight(45)
-        self.setEnabled(enabled)
-        self.clicked.connect(run)
-
-class PlotFilterWidget(QWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.canvas = FigureCanvas()
-        self.canvas.figure.set_layout_engine('constrained')
-        self.ax = self.canvas.figure.add_subplot(111)
-        self.ax.set_axis_off()
-        image_template = np.random.sample((28, 28))
-        self.heatmap = self.ax.matshow(image_template, cmap='cmap_image')
-        self.layout = QGridLayout()
-        self.layout.addWidget(self.canvas)
-        self.setLayout(self.layout)
-        self.setFixedSize(PLOT_TEST_WIDTH, PLOT_TEST_HEIGHT)
 
 class PlotCostWorker(QObject):
     finished = pyqtSignal()
@@ -478,292 +457,6 @@ class TrainWidget(QWidget):
         actions_panel.train_panel.start_button.setEnabled(True)
         actions_panel.train_panel.pause_button.setEnabled(False)
         actions_panel.train_panel.end_button.setEnabled(False)
-
-class ImageWidget(QWidget):
-    def __init__(self, parent):
-        super().__init__(parent)
-
-        self.main_window = self.parent()
-        self.X = None
-        self.labels = None
-        self.slices = None
-        self.Images = None
-        self.predictions = None
-        self.iterator = 0
-        self.loop_counter = 0
-        self.display = PlotFilterWidget(parent=self)
-        self.display_layout = QVBoxLayout()
-        self.display_layout.addWidget(self.display)
-        self.timer = None
-
-        self.next_button = Button('next', self.show_next, parent=self)
-        self.next_fail_button = Button('next fail', self.show_next_fail, parent=self)
-        self.previous_button = Button('previous', self.show_previous, parent=self)
-        self.previous_fail_button = Button('previous fail', self.show_previous_fail, parent=self)
-
-        self.actions_layout = QGridLayout()
-        self.actions_layout.addWidget(self.previous_fail_button, 0, 0, 2, 2)
-        self.actions_layout.addWidget(self.previous_button,      0, 2, 2, 2)
-        self.actions_layout.addWidget(self.next_button,          0, 4, 2, 2)
-        self.actions_layout.addWidget(self.next_fail_button,     0, 6, 2, 2)
-
-        self.layout = QVBoxLayout()
-        self.layout.addLayout(self.display_layout)
-        self.layout.addLayout(self.actions_layout)
-        self.setLayout(self.layout)
-
-
-    def init_data(self):
-        print(f'function init_data not implemented in base class ImageWidget')
-
-    def set_images(self):
-        data = self.main_window.model.data
-        self.Images = data.images(self.X)
-
-    def set_predictions(self):
-        model = self.main_window.model
-        self.predictions = model.predict(self.X)
-
-    def get_index(self):
-        if self.Images is None:
-            self.set_images()
-        return self.iterator % len(self.Images)
-
-    def get_title(self):
-        index = self.get_index()
-        return (f'Image # {self.get_index() + 1} / {len(self.Images)}: '
-                f'Prediction: {self.predictions[index]} '
-                f'True value: {self.labels[index]}')
-
-    def current_is_correct(self):
-        return self.predictions[self.get_index()] == self.labels[self.get_index()]
-
-    def show_current(self):
-        if self.Images is None or self.predictions is None:
-            self.set_images()
-            self.set_predictions()
-        index = self.get_index()
-        self.display.heatmap.set_data(self.Images[index])
-        self.display.ax.set_title(self.get_title())
-        self.display.canvas.draw()
-
-    def show_next(self, check_timer=False):
-        if self.Images is None or self.predictions is None:
-            self.set_images()
-            self.set_predictions()
-        self.iterator += 1
-        index = self.get_index()
-        self.display.heatmap.set_data(self.Images[index])
-        self.display.ax.set_title(self.get_title())
-        self.display.canvas.draw()
-
-        if check_timer:
-            self.loop_counter += 1
-            if self.loop_counter == len(self.Images):
-                self.timer.stop()
-                print(f'no failures were found')
-            elif not self.current_is_correct():
-                self.timer.stop()
-
-    def show_previous(self, check_timer=False):
-        self.iterator -= 2
-        self.show_next(check_timer)
-
-    def show_next_fail(self):
-        self.loop_counter = 0
-        self.timer = QTimer()
-        self.timer.timeout.connect(lambda: self.show_next(check_timer=True))
-        self.timer.start(1)
-
-    def show_previous_fail(self):
-        self.loop_counter = 0
-        self.timer = QTimer()
-        self.timer.timeout.connect(lambda: self.show_previous(check_timer=True))
-        self.timer.start(1)
-
-    def reset_predicted(self):
-        self.predictions = None
-
-    def test_extended(self):
-        print(f'function test_extended in base class ImageWidget not implemented')
-
-class ImageTestWidget(ImageWidget):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.actions_layout.removeWidget(self.previous_fail_button)
-        self.actions_layout.removeWidget(self.next_fail_button)
-        self.actions_layout.addWidget(QWidget(), 0, 0, 2, 2)
-        self.actions_layout.addWidget(QWidget(), 0, 6, 2, 2)
-        del self.previous_fail_button
-        del self.next_fail_button
-        del self.labels
-        self.init_data()
-
-    def init_data(self):
-        self.Images = None
-        slices = self.main_window.model.data.slice_test
-        self.X = self.main_window.model.data.X_test[slices]
-        self.show_current()
-
-    def get_title(self):
-        index = self.get_index()
-        return (f'Image # {self.get_index() + 1} / {len(self.Images)}: '
-                f'Prediction: {self.predictions[index]}')
-
-class ImageTrainWidget(ImageWidget):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.init_data()
-
-    def init_data(self):
-        self.Images = None
-        slices = self.main_window.model.data.slice_train_flat
-        self.X = self.main_window.model.data.X[slices]
-        self.labels = self.main_window.model.data.labels[slices[1]]
-        self.show_current()
-
-class ImageValidationWidget(ImageWidget):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.init_data()
-
-    def init_data(self):
-        self.Images = None
-        slices = self.main_window.model.data.slice_validation
-        self.X = self.main_window.model.data.X[slices]
-        self.labels = self.main_window.model.data.labels[slices[1]]
-        self.show_current()
-
-class FirstLayerMatrixWidget(QWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.main_window = self.parent()
-        self.matrix = None
-        self.figure =Figure(figsize=(20,20))
-        self.canvas = FigureCanvas(self.figure)
-        self.figure.set_layout_engine('constrained')
-        self.axes = None
-        self.quadmesh = None
-        self.subfigs = None
-        self.n_of_subfigs = None
-        self.n_of_filters = None
-        self.n_of_rows = None
-        self.quadmesh_template = None
-        self.layer_shape = None
-        self.layout = QGridLayout()
-        self.layout.addWidget(self.canvas)
-        self.setLayout(self.layout)
-        self.init()
-
-    def set_layer_shape(self):
-        config = self.main_window.config
-        nodes = self.main_window.model.grids_train._Grids__mknodes(config).astype(list)
-        shape = (nodes[1], nodes[0] + 1)
-        self.layer_shape = shape
-
-    def init(self):
-        self.set_layer_shape()
-        self.n_of_subfigs = 2
-        self.n_of_filters = self.layer_shape[0]
-        self.n_of_rows = -(self.n_of_filters // -self.n_of_subfigs)
-        self.canvas.setFixedSize(WEIGHT_FILTER_WIDTH * self.n_of_subfigs, WEIGHT_FILTER_HEIGHT * self.n_of_rows)
-        self.canvas.figure.clear()
-        self.subfigs = self.canvas.figure.subfigures(1, self.n_of_subfigs)
-        self.setup_axes()
-
-    def setup_axes(self):
-        self.axes = np.array([])
-        self.quadmesh = []
-        delta = 0.000001
-        count = 0
-
-        for subfig in self.subfigs:
-            self.axes = np.append(self.axes, subfig.subplots(self.n_of_rows, 1))
-        self.axes.flatten()
-
-        for ax in self.axes:
-            count += 1
-            self.quadmesh_template = np.random.sample(self.layer_shape[1] + 27 ).reshape(28, 29)
-            maximum = np.max(np.abs(self.quadmesh_template[:,1:]))
-            quadmesh = ax.pcolormesh(self.quadmesh_template, cmap='cmap_weights', vmin = -maximum - delta, vmax = maximum + delta, linewidth=0, rasterized=True)
-            ax.yaxis.set_inverted(True)
-            ax.set_axis_off()
-            ax.set_title(f'({count})', size=10, loc='left', y=1, pad=-10)
-            self.quadmesh.append(quadmesh)
-        self.zero_quadmesh()
-
-    def set_matrix(self):
-        print(f'set_matrix not implemented in base class FirstLayerMatrixWidget')
-
-
-    def get_filter_array(self, w_row):
-        self.set_matrix()
-        weights = w_row[1:]
-        bias = w_row[0]
-        filter_array = np.zeros_like(self.quadmesh_template)
-        filter_array[:,1:] = weights.reshape(28, 28)
-        filter_array[:,0] = bias
-        return filter_array
-
-    def zero_quadmesh(self):
-        for i in range(self.n_of_subfigs * self.n_of_rows):
-            array = np.zeros((28, 29))
-            self.quadmesh[i].set_array(array)
-        self.canvas.draw_idle()
-
-    def update_quadmesh(self):
-        self.set_matrix()
-        delta = 0.000001
-        vmax = np.max(np.abs(self.matrix)) + delta
-        vmin = -vmax
-        for i in range(len(self.matrix)):
-            array = self.get_filter_array(self.matrix[i,:])
-            vmax_i = np.max(np.abs(array))
-            vmin_i = vmax_i
-
-            vmax_i = 0.1
-            vmin_i = -vmax_i
-            self.quadmesh[i].set_clim(vmin, vmax)
-            self.quadmesh[i].set_array(array)
-        self.canvas.draw_idle()
-
-class FirstLayerWeightsWidget(FirstLayerMatrixWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.figure.suptitle(f'Weights (node)', color=mpl.rcParams['axes.titlecolor'], size=11)
-
-    def init(self):
-        super().init()
-        self.figure.suptitle(f'Weights (node)', color=mpl.rcParams['axes.titlecolor'], size=11)
-
-    def set_matrix(self):
-        self.matrix = self.main_window.model.grids_train.W[0]
-
-class FirstLayerGradientWidget(FirstLayerMatrixWidget):
-    def __init__(self, parent=None):
-        super().__init__(parent)
-        self.figure.suptitle(f'Gradients (node)', color=mpl.rcParams['axes.titlecolor'], size=11)
-
-    def init(self):
-        super().init()
-        self.figure.suptitle(f'Gradients (node)', color=mpl.rcParams['axes.titlecolor'], size=11)
-
-
-    def set_matrix(self):
-        self.matrix = self.main_window.model.grids_train.Gradient_W[0]
-
-class InspectWidget(QWidget):
-    def __init__(self, parent):
-        super().__init__(parent)
-        self.main_window = self.parent().parent().parent()
-        self.label = QLabel('INSPECT', alignment = Qt.AlignCenter)
-        self.layout = QGridLayout()
-        self.start_button = Button('weights', lambda: print(f'not implemented'), parent=self)
-        self.next_button = Button('gradient', lambda: print(f'not implemented'), parent=self)
-        self.layout.addWidget(self.label, 0, 0, 1, 2)
-        self.layout.addWidget(self.start_button , 1, 0)
-        self.layout.addWidget(self.next_button , 1, 1)
-        self.setLayout(self.layout)
 
 class ActionsWidget(QWidget):
     def __init__(self, parent):
