+++ ui/diagnostics_panel/predict_training_set/predict_training_set_view.py
@@ -1,0 +1,16 @@
+from PyQt5.QtWidgets import QWidget
+
+from ui.abstract.base_image import ImageWidget
+
+
+class ImageTrainWidget(ImageWidget):
+    def __init__(self, parent):
+        super().__init__(parent)
+        self.init_data()
+
+    def init_data(self):
+        self.Images = None
+        slices = self.main_window.model.data.slice_train_flat
+        self.X = self.main_window.model.data.X[slices]
+        self.labels = self.main_window.model.data.labels[slices[1]]
+        self.show_current()
